Edit Pool Button
